test(cli): split label-audit coverage
refs #272
